Seperate net syscalls
