+++ io_wrap.h
@@ -23,11 +23,6 @@ uintptr_t io_syscall_lseek(int fd, off_t offset, int whence);
 uintptr_t io_syscall_ftruncate(int fd, off_t length);
 uintptr_t io_syscall_pipe(int *fds);
 uintptr_t io_syscall_epoll_create(int size); 
-uintptr_t io_syscall_socket(int domain, int type, int protocol);
-uintptr_t io_syscall_setsockopt(int socket, int level, int option_name, const void *option_value, socklen_t option_len); 
-uintptr_t io_syscall_bind (int sockfd, uintptr_t addr, socklen_t addrlen);
-uintptr_t io_syscall_listen(int sockfd, int backlog);
-uintptr_t io_syscall_accept(int sockfd, uintptr_t addr, uintptr_t addrlen);
 uintptr_t io_syscall_fcntl(int fd, int cmd, uintptr_t arg);
 uintptr_t io_syscall_epoll_ctl(int epfd, int op, int fd, uintptr_t event);
 uintptr_t io_syscall_getcwd(char* buf, size_t size);
